+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 1.0.4
+ - Array of Pull API Raw data report columns
+ - method for getting mapped raw data report as array
+
 # 1.0.0
  - Pull API get raw report data method